Created DQC Rule 13, moved functions from rule 15 into neg_num.py, and added a function to facts.p
XS-2891
