+++ dqc_us_rules/util/facts.py
@@ -408,3 +408,22 @@ def grab_numeric_facts(facts_list):
         except ValueError:
             continue
     return numeric_facts
+
+
+def precondition_fact_exists(facts_list, precondition_concept):
+    """
+    Given a list of facts, check that the precondition concept is
+    included in the list
+
+    :param facts_list: list of facts to check for precondition concept
+    :type facts_list: list [:class:'~arelle.ModelInstanceObject.ModelFact'
+    :param precondition_concept: name of the concept to look for
+    :type precondition_concept: str
+    :return: return true or false depending on whether the precondition
+        concept exists
+    :rtype: bool
+    """
+    for fact in facts_list:
+        if (fact.concept.qname.localName == precondition_concept) and float(fact.value) > 0:
+            return True
+    return False
